Private Documentation

Documentation for VortexFilaments.jl's private interface.

See the Internals section of the manual for internal package docs covering all submodules.

Contents

Private Documentation

Contents

Index

Private Interface

Index

Base.isinf

Base.isinf

VortexFilaments.infdir

VortexFilaments.infdir

Private Interface

Base.isinf — Method

isinf(s::StaticArrays.SArray{Tuple{2},AbstractArray,1,2}) -> Any

Checks if s is an infinite segment.

Base.isinf — Method

isinf(vf::VortexFilament) -> Bool

Checks if vf is an infinite vortex filament.

VortexFilaments.infdir — Method

infdir(s::StaticArrays.SArray{Tuple{2},AbstractArray,1,2}) -> Any

Returns the index of the coordinate for which the first and/or second vertex of the segment s has an infinite value if it has one.

VortexFilaments.infdir — Method

infdir(vf::VortexFilament) -> Any

Returns the index of the coordinate for which the first and/or last vertex of the vortex filament vf has an infinite value if it has one.