Public Documentation

Documentation for VortexFilaments.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

Contents

Public Documentation

Contents

Index

Public Interface

Index

VortexFilaments.Vertex

VortexFilaments.VortexFilament

VortexFilaments.add!

VortexFilaments.getfreevertices

VortexFilaments.inducevelocity

VortexFilaments.inducevelocity

Public Interface

VortexFilaments.Vertex — Type

mutable struct Vertex <: StaticArrays.FieldVector{3,Float64}

Defines a vertex with at position (x,y,z).

Fields

x::Float64

y::Float64

z::Float64

VortexFilaments.VortexFilament — Type

struct VortexFilament

Defines a vortex filament of strength Γ consisting segments that connect the N vertices in vertices. Each segment in segments connects two subsequent entries of vertices. In case the first or last vertex has a coordinate value at infinity (Inf), there will be N-1 segments. Otherwise, there will be an N-th segment connecting the last and first vertex and thereby closing the filament. A vertex can be marked as a bound to a solid (e.g. a wing) if its index appears in boundidx. Similarly, a vortex is marked as a free vertex if its index appears in freeidx. If either the first or last vertex lies at infinity and the other does not, that other vertex is a bound vertex.

Fields

Γ::Real

Γ: Strength of the vortex filament.

vertices::Array{Vertex,1}

vertices: Array of vertices of the vortex filament.

segments::Array{StaticArrays.SArray{Tuple{2},Vertex,1,2},1}

segments: Array of segments of the vortex filament.

freeidx::Array{Int64,1}

freeidx: Array of indices of the free vertices of the vortex filament.

boundidx::Array{Int64,1}

boundidx: Array of indices of the bound vertices of the vortex filament.

VortexFilaments.add! — Method

add!(v::Vertex, Δv::Any) -> Any

Adds a vector Δv to the position of vertex v.

VortexFilaments.getfreevertices — Method

getfreevertices(vf::VortexFilament) -> Array{Vertex,1}

Returns the vertices of the vortex filament vf that are free.

VortexFilaments.inducevelocity — Method

inducevelocity(s::StaticArrays.SArray{Tuple{2},Vertex,1,2}, xeval::Any) -> Any

Computes the induced velocity by one segment s of a unit-strength vortex filament on the evaluation point xeval.

VortexFilaments.inducevelocity — Method

inducevelocity(vf::VortexFilament, xeval::Any) -> Any

Computes the induced velocity by the vortex filament vf on the evaluation point xeval.