Public Documentation

Documentation for VortexFilaments.jl's public interface.

See the Internals section of the manual for internal package docs covering all submodules.

Contents

Public Documentation

Contents

Index

Public Interface

Index

VortexFilaments.VortexFilament

VortexFilaments.VortexFilament

VortexFilaments.inducevelocity

VortexFilaments.inducevelocity

VortexFilaments.issemiinf

VortexFilaments.issemiinf

Public Interface

VortexFilaments.VortexFilament — Type

struct VortexFilament{TV<:(AbstractArray{#s12,1} where #s12<:(AbstractArray{T,1} where T)), TI<:AbstractArray{Int64,1}}

Defines a vortex filament of strength Γ, discretized by vertices in vertices and segments in segments, which connect the vertices. Each segment in segments should connect two subsequent entries of vertices. A vertex can be marked as a bound (e.g. to a wing) if its index appears in boundidx. Similarly, a vortex is marked as a free vertex if its index appears in freeidx.

Fields

Γ

Γ: Strength of the vortex filament.

vertices

vertices: Array of vertices of the vortex filament.

segments

segments: Array of segments of the vortex filament.

freeidx

freeidx: Array of indices of the free vertices of the vortex filament.

boundidx

boundidx: Array of indices of the bound vertices of the vortex filament.

VortexFilaments.VortexFilament — Method

VortexFilament(Γ, v)
VortexFilament(Γ, v, boundidx; isclosed)

Constructs a VortexFilament with strength Γ and vertices v. boundidx can be used to provide the indices of vertices that are considered as 'bound'. The constructor then set freeidx of the VortexFilament as the complementary list of indices. If the keyword isclosed is true, the segments of the VortexFilament will contain a segment that connects the first and last vertex in v, provided that these do not contain an infinite coordinate.

VortexFilaments.inducevelocity — Method

inducevelocity(s::StaticArrays.SArray{Tuple{2},AbstractArray,1,2}, xeval::Any) -> Any

Computes the induced velocity by one segment s of a unit-strength vortex filament on the evaluation point xeval. In case both vertices of the segment have an infinite coordinate (which have to be in the same direction), then the non-infinite coordinates are averaged over the two vertices to calculate the distance from xeval to the segment.

VortexFilaments.inducevelocity — Method

inducevelocity(vf::VortexFilament, xeval::Any) -> Any

Computes the induced velocity by the vortex filament vf on the evaluation point xeval.

VortexFilaments.issemiinf — Method

issemiinf(s::StaticArrays.SArray{Tuple{2},AbstractArray,1,2}) -> Any

Checks if s is a semi-infinite segment.

VortexFilaments.issemiinf — Method

issemiinf(vf::VortexFilament) -> Bool

Checks if vf is a semi-infinite vortex filament.